[clang] Disable synthesizing dealloc code for Objective-C classes by default
Summary:
We were creating very large code in these deallocs which isn't really needed for any of the current checkers so I'm disabling it to gain perf and avoid potentially buggy code. I'm leaving it there with a disabled flag in case we want to look into it again in the future.

Also, when calling dealloc, makes the code simpler: just call the user defined deallocs and do not try to deal with retain count.

Reviewed By: ngorogiannis

Differential Revision:
D58240612

Privacy Context Container: L1258973

fbshipit-source-id: df2b7ce077ace0f8e6fd68adf9d34ec9e4da34b0