[infer] tweak solver to avoid looping
Summary: The solver has been observed looping on some Kotlin and Erlang files. As far as I can tell this only occurs when something else has caused the formula to get corrupted (it has a cyclic formula of the form `v = instanceof(v,t)`, either because of erroneous recycling of variables or (less likely) miscompilation of the source). As a quick fix, we revert the recent replacement of `v>0` with `v != 0` in `propagate_atom` as that is observed to fix the problem for a problematic Kotlin file and only introduces one small regression in a Hack test. Further investigation is warranted, though.

Reviewed By: geralt-encore

Differential Revision:
D59104136

Privacy Context Container: L1208441

fbshipit-source-id: 5f223af91c8c696ee9dc7c543a3304147767160e