[self in block] Adding a new issue type for flagging when self is used in blocks that are passed to initialisers
Summary: This is a heuristic because often this leads to retain cycles given that the initialiser code will retain the block. If needed we can improve the checker in the future by checking that the second part actually holds, but just like this the check is intra procedural, so it remains simple.

Reviewed By: ngorogiannis

Differential Revision: D59281830

fbshipit-source-id: 7a38db1a798487912dfb0f51257f565136f92af7