Introduce the global/unified Lineage graph
Summary:
The unified Lineage graph is a whole-program graph, made from the per-procedure local graphs, that
holds the full interprocedural information (thus does not use interprocedural zipping nodes such as
`ArgumentOf` and `ReturnOf`).

It can only be generated when we have the final shape information for all procedures, to transform
those `ArgumentOf` and `ReturnOf` nodes into the proper `Argument` and `Return` with the correct
fields.

We then switch the DFS traversal in Lineage Taint to build a unified graph by doing that zipping node
transformation. The co-reachability is then much simplified as it can be implemented as a single
co-dfsin that graph.

We use the opportunity to directly export the result as a dot file, which circumvents the need to use
local scripts. Json export could still be added later if required.

Reviewed By: rgrig

Differential Revision: D59286583

fbshipit-source-id: b46460a436d1653820768038b2ba1a8dc706b698