[pulse] Fix ValueHistory.t consturction to satisfy the invariants
Summary:
This diff fixes the ValueHistory.t consturction to satisfy the invariants:

* `InContext` always comes first in a given history, if any, then `FromCellIds`, if any.
* They can appear at most once.

Before this diff, `InContext` appeared several times in a value history, thus it had many
`context`s, even with duplications in it, which resulted in inefficient runs of
`pop_least_timestamp`.

To fix the issue this diff revised the constructors to care about `InContext` and `FromCellIds`.

Reviewed By: jvillard

Differential Revision:
D59295293

Privacy Context Container: L1208441

fbshipit-source-id: b1603b72aa2e3092348181a9b5bc3d7af94c9391