Change --lineage-limit option to (an approximation of) the number of edges
Summary:
`--lineage-limit` allows to inspect the result of Lineage Taint requests which generate huge graphs
by limiting the visual output to a subset of the large graph.

It used to be interpreted as a limitation of the number of successor edges traversed in the
DFS. While this was very easy to implement and led to quick first experiments, this proved to not
scale out very well as, on real codebases, there was an enormous difference with limits of 3 or
4 (the former being uncomfortably small while the latter was completely out of reach of any human
inspection).

We change the interpretation of the limit into (something close to) the maximal number of edges in
the final graph. It then becomes much easier to finely tune it to get a result small enough to
visualise yet large enough that it contains relevant flows.

The current implementation works after reachability and co-reachability have fully run (without being
limited themselves). Then it repetitively adds the shortest path from the source to the sink into the
final result. To avoid adding the same path over and over again, when a path is added, its edges are
removed from the input graph.

Note that the final result may lose some proper graph structures. For instance, on a graph including
diamonds such as `Source -> {A, B} -> C -> Sink`, since the edge `C -> Sink` will be deleted on the
first try, `A` and `B` cannot appear both as they both require this edge to flow into the
sink. A more sensible implementation is future work.

Reviewed By: rgrig

Differential Revision: D59288762

fbshipit-source-id: 0a28869d0b76737786b4144f3e11ffb04bc3d8ec