Make map stdlib functions appear in Lineage graph
Summary:
The core of the precise map support on atom-keys was to interpret the map creation and access
functions as if they were precisely accessing fields of a structure, hence with specific models that
performed something similar to a field lookup operation.

Yet it appeared that just doing that hadf the unfortunate effect of making these function names
disappear from the graph, in particular as Argument and Return nodes.

We keep the same precise interpretation of these functions but explicitly make their argument and
return nodes appear in the graph with the corresponding call and return edges.

The "precise map support" models can now be seen as generating "Derive" edges more precise than the
standard one (which do not specifically consider maps at all) while keeping the Call and Return
present to be able to query them in Lineage Taint.

Reviewed By: rgrig

Differential Revision: D59289834

fbshipit-source-id: e7a592447f1a96e799818909d49d18f5611ee2a8