[hack][reactive] fix tenv merging
Summary:
Type environment merging merges only the explicitly captured dependencies. This means that in reactive capture, or when continuing a capture, the old type environment is thrown away and not merged into the result.

This diff fixes this by making the merge start from the existing global type environment when doing reactive capture. We also forcibly reload the global type environment after the merge, since the merge happens in a forked process and thus will not affect the global reference in the parent process.

Reviewed By: davidpichardie

Differential Revision:
D59368827

Privacy Context Container: L1258973

fbshipit-source-id: 7c38ffc3cd839dfc5bd5e519c24f6ba77a815961