[merge] fix source file handling
Summary:
The addition of `source_file` to `Struct.t` failed to update the merge process for that field, meaning that currently the field gets a non-deterministic value if the same struct is seen twice, but with different values for `source_file`.

This diff also removes the use of `with` syntax so as to force future additions to the `Struct.t` type to explicitly change the `merge` function.

Reviewed By: thizanne

Differential Revision:
D59639898

Privacy Context Container: L1258973

fbshipit-source-id: 9ec876b466771e6b8c7affc652a15d36a58c0c2e