[pulse disjunctive domain] less depth-first exploration of instruction blocks
Summary:
When a node contain several instructions Pulse was first executing all instructions from the first disjunctive precondition, then, if room left, from the 2nd...

Now we wait for the full analysis of the first instruction to jump to the 2nd. We expect to analyse more paths with this approach.

Reviewed By: skcho

Differential Revision:
D55798800

Privacy Context Container: L1208441

fbshipit-source-id: e057cd8fdf85dcb2f096dcd58a53ce76d372ecea