[absint] specify current node when reporting
Summary:
The heuristic to detect file renamings is only activated if the bug's
error description includes a `node_key`, which is a "hash" of the CFG
node where the issue was found. Before this diff, this node was only set
from biabduction. Now it should be set anywhere the current node is
registered, which is most places. The exception would be checkers that
report issues outside of `exec_instr` without opening a new session (for
debug html, for example). I haven't audited this but Pulse is in the
clear there.

Reviewed By: ngorogiannis

Differential Revision:
D59369325

Privacy Context Container: L1258973

fbshipit-source-id: bb6081dc14aab096ebb6b13b1eb737480c4471b7