Make `__erlang_make_map` appear in Lineage graph, allow it as a source/sink
Summary:
Contrary to functions such as `maps:put`, which are standard procedures and could be added easily to
the Lineage graph to allow querying them (D59289834), `__erlang_make_map` is not a standard procedure
but an internal builtin that is used to encode the `#{ K => V}` syntax.

As it still appears useful as a queryable source/sanitizer/sink (eg. to answer a question like "what
could go in a map?"), we have it appear in the Lineage graph the same way as `maps:put` and similar.

To allow taint queries involving it, we add to the standard mfa parsable arguments an explicit list
of allowed internal builtins with hard-coded names (which ideally will match their internal name for
more clarity). That list currently only contains `__erlang_make_map` but could be extended on
a case-by-case basis (we definitely do not want every builtin to appear here as the usual approach is
to actually try and remove them to reduce noise).

Reviewed By: rgrig

Differential Revision: D59818127

fbshipit-source-id: 690f2b8bbf1d1161f44126d3c32da37cae93d666