Introduce Instrumented graph module in Lineage Taint
Summary:
Some Lineage taint algorithms involve instrumenting the graph with single source and sink nodes, in
particular to have an easy way to find "the shortest path from any source."

As this is going to have more use cases, we introduce a separate abstract type (in a module) to
represent this graph and avoid mixing it with the uninstrumented one. This makes it compiler-checked
that we don't use uninstrumented graphs when we need the instrumentation, and that the
instrumentation doesn't appear visible to the end user.

Reviewed By: rgrig

Differential Revision: D59917464

fbshipit-source-id: c6aaf4b42fd264e8dbce55b58950bb60cae758b3