Generate traces in LineageTaint
Summary:
We use the taint information computed by Lineage to generate traces that exhibit data flow from
source to sink.

Let's call "sinking edges" the edges whose destination is a sink node. For instance, if the sink is
an argument of a function, the sinkind edge would be the Call edges into that function (for that
argument).

This diff will generate one trace per sinking edge, by finding out the shortest path from any source
to the source of the edge. Those traces will be put in the json format that traceview understands,
then the resulting json file can be uploaded to manifold (with a chosen unique name) and the
traceview link will become immediately accessible.

Reviewed By: rgrig

Differential Revision: D59965688

fbshipit-source-id: 244f9b7972deca91702edb8c0f54408b5cca5331