[clang] print source file once it is known
Summary:
This is for the benefit of the parallel clang capture: when we start the
capture we don't know the name of the source file being captured, but we
learn it later on. Sadly clang takes a long time to tell us so in order
to display the correct time we also modify the API of the progress bar
to allow changing the title of the task without resetting the timer of
the task.

With this change, we can now start by displaying "capturing source file"
and then update the string to be the name of the file.

Reviewed By: martintrojer

Differential Revision:
D61132934

Privacy Context Container: L1208441

fbshipit-source-id: e782643b5365552de78bb455d3e63121947ff183