[pulse] cleanup variables using `always_reachable`
Summary:
For some reason the extra dereference is needed for objc but in C
programs it is incorrect and caused NPEs in the test:

```
__attribute__((cleanup(cleanup_int))) int* x = malloc(...);
```

We ended up checking the attribute on `*x` instead of `x`, so in the
NULL case of malloc() it was an NPE, and in the non-NULL, allocated
case, it was removing the attribute on `*x` instead of `x` so we still
had a leak.

Instead, opt for a different approach that considers `x` always
reachable so that it doesn't need to be dereferenced. The advantage is
that the approach is uniform and is equivalent to a "deep" deallocation.
The incovenient is that objects stick around and are forever marked as
allocated and reachable (until they disappear from the state), but this
is minor.

(Ideally we should try to call the cleanup method!)

Fixes https://github.com/facebook/infer/issues/8 at long last

Reviewed By: geralt-encore

Differential Revision:
D61115436

Privacy Context Container: L1208441

fbshipit-source-id: 8216d896133f539b71d7d68193e891680c0cda19