[pulse] manifest if conditions only of depth 0
Summary:
Latent issues prevent us from reporting issues that are dependent on
some contextual information that might never realise. This is important
for inter-procedural analyses because a given function might call other
functions that have different expectations of what is safe, and
generally expect their own invariants to hold. However, this should not
prevent us from reporting issues whenever it is the same function making
assumptions and then having incorrect behaviours under those
assumptions.

Concretely:

```
foo(int x) {
  if (x > 5) {
    int* p = NULL;
    *p = 42; // should report!
  }
}

int* null_if_zero(int x);

// expects x ≠ 0
void goo(int x) {
  int* p = null_if_zero(x);
  *p = 42; // latent issue
}
```

We want to report on foo(): x > 5 is seemingly a possible input since
the function itself accounts for it, yet it is unsafe if x > 5. The
reponsibility for the bug is very likely in foo(), not callers of foo()
that call it with a parameter greater than 5.

We don't want to report on goo() in general as it's possible that its
contract is that x must be > 0. The fact that a callee can return NULL
if passed x=0 might be because the callee is extra cautious compared to
goo() so to avoid FPs it is better to delay reporting until we see a
goo(0) call.

TODO: This doesn't yet solve `if (p == NULL) { *p = 42; }`, where we
would also want to report an error. This is because the error will be
filtered out because `p` has never been actually assigned to NULL.

Reviewed By: skcho

Differential Revision:
D60970323

Privacy Context Container: L1208441

fbshipit-source-id: 90e28ea36a7fff3d179c7a13a6eff3de8b28f408