[pulse] add and report COMPARED_TO_NULL_AND_DEREFERENCED
Summary:
The goal is to emit an error for the classic blunder `if (p == NULL) { *p = 42; }`.

Since `p` is never actually assigned `NULL` here, pulse won't report. We
want to tell pulse it's ok to report if the comparison to NULL happened
in the same procedure as the dereference. We do that with an
intra-procedural attribute to record the comparison to NULL.

Right now this creates false positives because we should check that we
are dereferencing exactly the expression that was compared to NULL, not
just anything that happens to be equal to 0.

Reviewed By: geralt-encore

Differential Revision:
D60970325

Privacy Context Container: L1208441

fbshipit-source-id: ea945ed4ba98a4744fe879c5b20d9cdfdc9632b5