[pulse] add invalidation events to value origin
Summary:
When we prune `p == NULL` we want to record that `p` was invalidated.
However, since we actually do

```
n$0 = *&p
PRUNE(n$0)
```

for now we will record that `n$0` was invalidated. The following diffs
make it so that we can propagate this back to `p` too.

This will allow us to restore the heuristic that checks that a value has
been invalidated before reporting null-like issues.

Reviewed By: geralt-encore

Differential Revision:
D60970321

Privacy Context Container: L1208441

fbshipit-source-id: d7ee95a16c63acdce242f3534aa6deb815edaa0a