[pulse] stacks hold value origins now
Summary:
Logical variables `n$0` etc. are used to hold values from the memory and
from program variables. It is useful to remember what was loaded into
these logical variables so as to be able to record observations about
their original value later on. So, store value origins instead of
"addr,hist" pairs in the stack image. The origin is always Unknown for
now, which is the same as an "addr,hist" pair.

Reviewed By: skcho

Differential Revision:
D60970317

Privacy Context Container: L1208441

fbshipit-source-id: 34b6757625404b5a003fe2e3671ba3e57aaf48a8