[pulse] record value origin on load
Summary:
When we prune `p == NULL` we actually do:

```
n$0 = *&p
PRUNE(n$0)
```

so we need to remember that `n$0` is holding the value of `p` and
record the invalidation event on `*&p` as well as `n$0`.

This will allow us to restore the heuristic that checks that a value has
been invalidated before reporting null-like issues.

Reviewed By: skcho

Differential Revision:
D60970319

Privacy Context Container: L1208441

fbshipit-source-id: 13e5b8cbe2f9c23eef873c8b8d1e2c25c2b4da25