[pulse] be more cautious about matching trace invalidations to reports
Summary:
We were doing an optimistic, unsound filtering before, where any
invalidation event in the trace would be good enough to justify
reporting a null pointer dereference (or comparison to null). Now that
we record comparisons to nulls as invalidations this can introduce false
positives so fix the filtering to look for the *same* invalidation
event than the one we are trying to justify.

Reviewed By: geralt-encore

Differential Revision:
D61202960

Privacy Context Container: L1208441

fbshipit-source-id: 09b9679ac2e78c3903419890edc9dea257d63c34