[IR] Refactor the data tructures for captured variables
Summary:
The type of the captured variables when they were part of a closure was a tuple, however the same information was stored as a record for the captured variables that go in the procedure attributes. This diff unifies the data types.

This is with the goal of adding another element to the captured variables `is_formal` to check if it's a parameter of the parent method.

Reviewed By: skcho

Differential Revision: D61331549

fbshipit-source-id: eac548163044e0ccf8497ebb21c2e53e62cd7f4e