[unnecessary copy] Revise the logic for determining used-after-copy
Summary:
This diff revises the logic for determining whether a variable is used after a copy.  The logic
depended on a *global* timestamp comparison in non-disj domain, but the timestamps are bound to each
disjunct, so which introduced some FN.  For example,

```
for(...) {
  ...
  foo(x); // (1) x is used
}

y = x; // (2) x is copied
```

suppose `x` is used in a loop and copied after that.

* In a disjunct that exited the loop quickly, it may have a small timestamp at (2).
* In a disjunct that stayed in the loop longer, it may have a bigger timestamp at (1).

Thus, the reasoning at non-disj domain may conclude "`x` may be used after the copy" incorrectly.

To avoid such a case, this diff adds a location comparison in the check as heuristics.

Also, added a dummy semantic model for `apache::thrift::field_ref::operator=`, so that it is not addressed as unknown that havoc's passed arguments.

Reviewed By: dulmarod

Differential Revision: D61274966

fbshipit-source-id: ad9a0dd1be7711a3dbe6175b14d8f6b14f1f9843