[annot] Add sink definition to trace
Summary:
Annotation reachability has two entries in the trace for each step: the callsite plus the definition of the callee. Except for the final step (the sink), which only had the callsite. This diff adds the sink definition too. This is helpful during debugging but I think developers will also find it useful.

Plus it makes the code cleaner: no need for the extra step to add the final callsite separately.

Reviewed By: ngorogiannis

Differential Revision:
D61852558

Privacy Context Container: L1208441

fbshipit-source-id: 43126c2d2f072b78c714492e28a593e11eebcc30