[pulse] add taint propagation to file operations C models
Summary:
Modelling functions has the drawback that it disables the default
heuristic for unknown functions and taint propagation. This can be
solved by either adding these models to a list in TaintOperations of
modelled functions for which we should still apply the heuristic, but
it's better to directly model the correct taint behaviour inside the
models, as is done here.

Reviewed By: ngorogiannis

Differential Revision:
D61850090

Privacy Context Container: L1208441

fbshipit-source-id: 8618912619ba5924bef3affb069aa42f93b3df9b