[pulse] recording of fine-grained callees
Summary:
We record call graph edges caller->callee but taking into account the specialization key used for the callee.
We rely on the preexisting PulseTransitiveInfo framework.

The purpose is to enhance all transitive informations in order to avoid storing a full transitive information in each summary.
We use this technique in the next diff.

Reviewed By: ngorogiannis

Differential Revision:
D61790708

Privacy Context Container: L1208441

fbshipit-source-id: 3f380b7007148981a297cf4238d7c589ff43822a