[pulse][DSL] defer model execution
Summary:
OCaml is annoying and will happily start evaluating models until it hits
something that waits for an explicit argument. Make that thing the first
step in models to prevent this in a bunch of cases.

This can cause actual programming errors where fresh abstract values are
generated once per model instead of on each invocation, although I
don't have good evidence of that.

Reviewed By: davidpichardie

Differential Revision: D61985925

fbshipit-source-id: 8f16efe416ed4247c8be3c5604ec39ce7938f82f