[analysis] remove ProcUID target
Summary: The scheduler target type `ProcUID` was added as an optimisation (instead of serialising a procname and sending it across a pipe). Testing shows it's not performing any better, and it enlarges the interface every scheduler needs to implement. This diffs replaces the uses of `ProcUID` with `Procname`.

Reviewed By: jvillard

Differential Revision:
D62016743

Privacy Context Container: L1208441

fbshipit-source-id: b32f382f247c704aa475ed8431ba5cac5465f8e7