[pulse] get rid of contexts
Summary:
Contexts are (currently) expensive to keep around and are only used by
the clang frontend. Their goal is to remember the histories of how we
passed conditional branches' tests, augmenting the traces we report.
This works kind of ok but you need to have structured programming to
get these contexts (need to know which branches you are currently
"under"). We'll eventually need some capability like this because it is
useful to report this context in traces but since this was limited to
clang (and even there not fully working) the solution should be
different. For example we could remember each path's set of passed
conditionals and show some or all of these when generating traces.

Anyhow, this deletes contexts and the associated ValueHistory machinery,
which was a bit complicated.

Reviewed By: skcho

Differential Revision: D62031136

fbshipit-source-id: 9db13f2e61f6f993e0eb9fee71988d709caa66a8