[cxx_ref] Take into account an external list of functions with no escaping block arguments
Summary:
There are cases where the annotation NS_NOESCAPE cannot be used because it's only used for blocks but in Objective-C++ we can use a templated argument to be instantiated with a block. For these cases we are adding an infer option to add those methods to infer.

This is to avoid false positives of CXX_REF_CAPTURED_IN_BLOCK.

Reviewed By: ngorogiannis

Differential Revision: D62180750

fbshipit-source-id: 083f95af4bda125296ef039520e5f6b710c52ec0