[erlang] Add more tests for matching against bound variables
Summary: Previously we had a test where we match a variable against a bound variable, which also requires a good equality check model. In this test, we match a constant against a bound variable so this only depends on us translating variable patterns the right way (which we don't do yet, so the test is false negative).

Reviewed By: rgrig, thizanne

Differential Revision: D62234243

fbshipit-source-id: f7bf21f7ed80eb4fb2a4eb41ee03c5e8e3c4c64e