[pulse][RFC] new disjunct selection strategy at call sites
Summary:
when executing an instruction on several disjuncts, we give to each disjunct a "budget" to not exceed. This makes the implementation of the "balanced" strategy far easier that what I was currently working on, with a risk that some disjunct will not use all their budget.

For example, if a caller has currently 3 disjuncts and calls a callee whose summary has 12 disjuncts, with max_disjunct=20:
- caller 1st disjunct will get a budget of 6+1
- caller 2nd disjunct will get a budget of 6+1
- caller 3rd disjunct will get a budget of 6

7+7+6 = 20

In the current diff we activate the new approach by default for Hack, in order to see the impact of the new strategy on smoke test. This is temporary.

A slight improvement is proposed in D62236165, but without big impact on smoke.

Reviewed By: nbenton

Differential Revision:
D61975381

Privacy Context Container: L1208441

fbshipit-source-id: 6d6805f66f99d7e8f12a9e040ff06306c3e02ce8