[clang] correct translation of bound references
Summary:
clang lies to our faces about the type of bindings at their point of
use:

```
  // m is a map int -> int
  auto &[k, v] = *m.begin(); // here v is a ref "int&", we can see it in the AST
  // actually, we can only see that the whole &[k, v] is a ref...

  int* vp = &v; // here v is suddenly just an int, lololol
  // it really is an int& though!
```

We can fix this by looking back at the decl ref corresponding to the
variable and check if the expression it binds is itself a ref. If so, we
add a `&` to the type we get from the AST. Disgusting.

I hope it works in more complex cases...

Reviewed By: ngorogiannis

Differential Revision: D62442049

fbshipit-source-id: d924273f0ec976c25f2362541e5a5d754fbd922b