[IR] Adding context info data to captured variables.
Summary: We want to add some context information to the captured variables. Here we add whether the variable has been checked for null before being captured, which is useful in some checkers. We will be adding a preanalysis to compute this data and add it to the captured variables in the attributes.

Reviewed By: ngorogiannis

Differential Revision: D62385176

fbshipit-source-id: 4764d3eeb72d736cc8e03c8846188b7e6d06eccb