[reporting] Adding optional autofix field to the reports
Summary: We add a new collection of fields to the report.json for specifying how to suggest fixes for the issues found. Currently we'll add original and replacement, which work together with the line and column of the issue type, that are also the location of the original input to be changed to the replacement. This will work for simple one line replacements, maybe more fields will be needed for more complex ones.

Reviewed By: martintrojer, ngorogiannis

Differential Revision: D62579331

fbshipit-source-id: a5fc7b2d5d067296396c3a8695540675c4b12db2