[annot] Detect if a callsite on the path is inside a loop
Summary:
Add a new option to check if any callsite on the path from source to sink is nested inside a loop. The idea is that such issues might be more severe. Currently, we just display this information in the trace, but we can also lift it to the issue's main description (or even introduce a new issue kind).

Remark: the check whether we are in a loop is done on the CFG using `Loop_controls`, which is a heuristic/approximation. But, for control flow graphs coming from "well-behaved" Java programs (no goto), I expect them to work reasonably well. An alternative could be to try `WTO` but that is also an approximation based on the CFG. The ultimate source of truth would probably best come from the frontend (AST).

Reviewed By: ngorogiannis

Differential Revision: D62633560

fbshipit-source-id: fd1bcc1dc5c1019f33bd0f4d12cb7ab9e49a9041