[annot] Fix bug: add entry instead of overwriting
Summary: Fix a bug in annotation reachability domain, where if we try to add a callsite to a sink, and the sink has not yet been seen, we simply overwrite existing sinks by putting in a new singleton map instead of adding to the existing map. Interestingly, this bug didn't (always) cause false negatives because we have exceptional edges leading to the procedure exit, where the join operation gets all the previously overwritten maps and makes a union.

Reviewed By: ngorogiannis

Differential Revision: D62736430

fbshipit-source-id: 2cd055256159ce48c0dc0bcf07ba2d0e18876e6f