[annot] Add option to minimize paths on sink side
Summary:
Previously, paths always stopped at the first sink, but on the source side it was configurable. Recent fixes now allow the sink side to be configured too. So given a `source1() -> source2() -> sink1() -> sink2()` chain, we can configure to start reporting from both `source1` and `source2`, or `source2` only, and on the sink side we can configure to stop at `sink1` or to report both `sink1` and `sink2`:

| Path | No minimization | Minimize source | Minimize sink | Minimize both |
| `source1() -> source2() -> sink1() -> sink2()` | X | | | |
| `source1() -> source2() -> sink1()` | X | | X | |
| `source2() -> sink1() -> sink2()` | X | X | | |
| `source2() -> sink1()` | X | X | X | X |

Reviewed By: ngorogiannis

Differential Revision: D62846218

fbshipit-source-id: 49b300e718e6566fe5200a6b725930473d4c1b6c