[inferpython] new tracking of qualified names
Summary:
each function in a Python module can be uniquely identified by a qualified name. In Python3.8 this name is not part of the code object but is inserted by the compiler at each call site. 3.10 directly attachs this qualified name to each code_object.

We build a similar map (than 3.10) that attaches a qualified name to each code object. We use this map at call site to register the right qualified name when generating function call instructions.

Reviewed By: ngorogiannis

Differential Revision:
D62855949

Privacy Context Container: L1208441

fbshipit-source-id: a3ec0789c09e37466ad165ac4245168229ad9f44