[pulse] Add pulse model to model handle failure functions to define assertions
Summary: Underneath of assert definitions, which are generally a macro, there is a `handle_failure` method that we need to model to handle assertions effectively. In pulse as well as in biabduction we deal with these functions as `assume false` because we don't want to report on those paths.  This adds a config flag to add other custom `handle_failure` functions.

Reviewed By: jvillard

Differential Revision: D63099687

fbshipit-source-id: 4daefb79710b0c8b441fd61e926c7ba19b39862b