[restart] re-entrant locks
Summary:
Allows locking the same proc name multiple times from within the same
process, and unlocking only at the outer-most locking site.

This is never triggered for now because ondemand will raise a
MutualRecursion exception before we ever get here, but we need this
behaviour for the top of the stack.

This isn't enough as this introduces livelocks due to ondemand and the
restart scheduler having different opinions on what a work item is.

Reviewed By: ngorogiannis

Differential Revision:
D63032163

Privacy Context Container: L1208441

fbshipit-source-id: 61a2981a51180406ce39854748a20e61aebfae07