[restart] send the list of dependencies on restart instead of the last failed lock
Summary:
The goal is to lock ahead of time all the live (i.e. with a summary yet to be
computed) dependencies of a procedure when we are restarting the analysis
of that procedure (well, of the corresponding work packet that led to the
analysis of that procedure). Start by sending what the current dependencies
were at the time of the restart race condition, which are those live dependencies.

Reviewed By: ngorogiannis

Differential Revision:
D63032161

Privacy Context Container: L1208441

fbshipit-source-id: 8f54535d67eeeb5f09357d532c0c0337cfd78462