[restart] do not dequeue until all deps are locked
Summary:
The diff that actually does something!

When we ask for a restart on a work packet and that work is getting
scheduled again, this time pre-emptively lock all the dependencies we
know we were exploring when we last started that work.

If that fails because another process is working on one of these
procedures, try again. Until we get all the locks. Since no other work
is getting scheduled in the meantime, eventually we'll have no other
process to contend with and should be able to get on with our lives.

Reviewed By: ngorogiannis

Differential Revision: D63032160

fbshipit-source-id: f03987667e4c93a8bb568d58670b308f8b7c7173