[processpool] stop checking if there is new work after the first "no"
Summary:
Checking for new work is now potentially expensive as we try locking a
bunch of files (corresponding to proc names) in the restart scheduler.
The process pool scheduler was asking for new work for each idle child
so if we are waiting for a file lock we would repeatedly try acquiring
a bunch of locks, which can be taxing for the system. Anecdotally I saw
the progress bar become a bit spastic at times where we are trying to
acquire several locks to unblock an analysis.

Now we stop at the first idle child that gets no work, as it's unlikely
that checking again immediately would give us a different answer. We
will check again at the next timeout or child message.

Reviewed By: thizanne

Differential Revision: D63094848

fbshipit-source-id: 4a252aa06adab4b8bdbf65357dc7fbc85dc9c0d2