[restart] do not reserve more than one job for a given ready child
Summary:
There was a problem with the previous approach: we would try to reserve
and move to the "ready" queue as many jobs as we managed to
pre-emptively lock. But, we locked these jobs for a single child's PID,
meaning when they got scheduled they would immediately fail unless they
are exactly that child. This still worked eventually but obviously it's
inefficient. Instead of preparing jobs in advance, prepare and send only
one job at a time. Wait until the first job in the blocked queue is
ready to go. This is surprisingly more efficient than trying to schedule
as many jobs from the blocked queue as we can.

Also added a few more comments about how this works.

Reviewed By: ngorogiannis

Differential Revision: D63271427

fbshipit-source-id: 096d6ad1d37d8be260ee2828f61b9ac7c82b0432