[ondemand] correctly detect mutual recursion when requesting specialisation in the absence of a default summary
Summary:
If we want to specialise a procedure that has no summary, we first
compute its "default" summary, then add the specialised one. This was
done by computing the default summary while the specialised summary is
registered, then recursively calling `analyze_callee`, which will try to
register the same procedure again. This is exactly wrong and must be the
other way around.

Reviewed By: ngorogiannis

Differential Revision: D63395290

fbshipit-source-id: 5c70a7643a111626eba6d2a265660fd31c504400