[inferhack] use more temp var when manipulating mutable expressions
Summary:
The Python frontend removes bytecode operand stack using
a standard technique: we execute symbolically the bytecode using a stack
of symbolic expression. But if we push an expression that contains symbols
whose value could be changed before we pop the expression, it could lead
to wrong translation. We now avoid this situation by using more temporary
variables. The analysis backend is likely to prefer this pattern anyway.

Reviewed By: ngorogiannis

Differential Revision:
D63031203

Privacy Context Container: L1208441

fbshipit-source-id: ef452a70f42f9b156a5e25ede5a0f4e6d635d887